Use Nextcloud provider even if no JSON was returned
Nextcloud API calls do not return JSON when Nextcloud is not set up. The new behaviour handles this in the Nextcloud provider similar as if it were unauthorized, making it possible to use the handler in this case. It is required to properly redirect to login form when Nextcloud is not set up.